+++ src/ump/api/routes/jobs.py
@@ -8,9 +8,9 @@ from flask import Response, g, request
 from sqlalchemy import create_engine, or_, select
 from sqlalchemy.orm import Session
 from ump import config
-from ump.api.ensemble import JobsUsers
-from ump.api.job import Job
-from ump.api.job_comments import JobComment
+from ump.api.models.ensemble import JobsUsers
+from ump.api.models.job import Job
+from ump.api.models.job_comments import JobComment
 from ump.api.jobs import append_ensemble_list, get_jobs
 from ump.api.keycloak_utils import find_user_id_by_email
